Add support for building Xbox360 builds
The Sega Vintage Collection[1] has two "extra" versions of the TJ&E ROM:

- ToeJamEarl_2012_06_21_MENU_FIX.GEN
- ToeJamArl0618_sawa.GEN

These are based on REV02. I haven't quite worked out what the "MENU FIX"
is yet, but the "sawa" version seems to correspond to the "Rocket Skates
Rhapsody" challenge mode (although it seems to require additional state
to be loaded).

Add disassemblies for both of these to this repo.

https://segaretro.org/Sega_Vintage_Collection:_ToeJam_%26_Earl